Stage 4/7: A good stack
Description
Imagine a situation: the answer is wrong for the given term, but it is correct for another term. Let's consider situations like this.

You can use maps. Ask all card's definitions in the order of addition. If the definition is wrong for the current term but it is correct for another, output the original term.

When the user tries to add a duplicated term or definition, forbid it and ask again until the user inputs a unique one. For now, you are able to implement this without a try catch construction. Use the rule: if you can avoid exception-based logic, avoid it!